Move SofaMessage to sofa dir
