+++ app/src/main/java/com/tokenbrowser/manager/store/PendingMessageStore.java
@@ -19,7 +19,7 @@ package com.tokenbrowser.manager.store;
 
 
 import com.tokenbrowser.model.local.PendingMessage;
-import com.tokenbrowser.model.local.SofaMessage;
+import com.tokenbrowser.model.sofa.SofaMessage;
 import com.tokenbrowser.model.local.User;
 import com.tokenbrowser.util.LogUtil;
