+++ app/src/main/java/com/tokenbrowser/model/local/Conversation.java
@@ -18,6 +18,8 @@
 package com.tokenbrowser.model.local;
 
 
+import com.tokenbrowser.model.sofa.SofaMessage;
+
 import java.util.List;
 
 import io.realm.RealmList;
