+++ app/src/main/java/com/tokenbrowser/model/sofa/SofaMessage.java
@@ -15,9 +15,10 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.tokenbrowser.model.local;
+package com.tokenbrowser.model.sofa;
 
-import com.tokenbrowser.model.sofa.SofaType;
+import com.tokenbrowser.model.local.SendState;
+import com.tokenbrowser.model.local.User;
 
 import java.util.UUID;
 import java.util.regex.Matcher;
